+++ src/Yasumi/Provider/AbstractProvider.php
@@ -22,6 +22,7 @@ use Yasumi\Filters\BetweenFilter;
 use Yasumi\Filters\OnFilter;
 use Yasumi\Holiday;
 use Yasumi\ProviderInterface;
+use Yasumi\SubstituteHoliday;
 use Yasumi\TranslationsInterface;
 use Yasumi\Yasumi;
 
@@ -302,13 +303,15 @@ abstract class AbstractProvider implements ProviderInterface, Countable, Iterato
      */
     public function count(): int
     {
-        $list = $this->getHolidayNames();
+        $names = \array_map(static function (&$holiday) {
+            if ($holiday instanceof SubstituteHoliday) {
+                return $holiday->substitutedHoliday->shortName;
+            }
 
-        \array_walk($list, static function (&$holiday) {
-            $holiday = \str_replace('substituteHoliday:', '', $holiday);
-        });
+            return $holiday->shortName;
+        }, $this->getHolidays());
 
-        return \count(\array_unique($list));
+        return \count(\array_unique($names));
     }
 
     /**