+++ tests/UnitedKingdom/NorthernIreland/SpringBankHolidayTest.php
@@ -1,4 +1,4 @@
-<?php declare(strict_types=1);
+<?php
 /**
  * This file is part of the Yasumi package.
  *
@@ -10,51 +10,49 @@
  * @author Sacha Telgenhof <me@sachatelgenhof.com>
  */
 
-namespace Yasumi\tests\Romania;
+namespace Yasumi\tests\UnitedKingdom\NorthernIreland;
 
 use DateTime;
 use DateTimeZone;
-use Exception;
-use ReflectionException;
 use Yasumi\Holiday;
 use Yasumi\tests\YasumiTestCaseInterface;
 
 /**
- * Class for testing Saint Andrew Day in Romania.
+ * Class for testing the Spring Bank Holiday in Northern Ireland.
  */
-class StAndrewDayTest extends RomaniaBaseTestCase implements YasumiTestCaseInterface
+class SpringBankHolidayTest extends NorthernIrelandBaseTestCase implements YasumiTestCaseInterface
 {
     /**
-     * The name of the holiday to be tested
+     * The name of the holiday
      */
-    public const HOLIDAY = 'stAndrewDay';
+    public const HOLIDAY = 'springBankHoliday';
 
     /**
      * The year in which the holiday was first established
      */
-    public const ESTABLISHMENT_YEAR = 2012;
+    public const ESTABLISHMENT_YEAR = 1965;
 
     /**
-     * Tests Saint Andrew Day on or after 2012.
-     * @throws Exception
-     * @throws ReflectionException
+     * Tests the holiday defined in this test.
+     * @throws \Exception
+     * @throws \ReflectionException
      */
-    public function testStAndrewDayOnAfter2012()
+    public function testHoliday()
     {
-        $year = $this->generateRandomYear(self::ESTABLISHMENT_YEAR);
+        $year = 1988;
         $this->assertHoliday(
             self::REGION,
             self::HOLIDAY,
             $year,
-            new DateTime("$year-11-30", new DateTimeZone(self::TIMEZONE))
+            new DateTime("$year-5-30", new DateTimeZone(self::TIMEZONE))
         );
     }
 
     /**
-     * Tests Saint Andrew before 2012.
-     * @throws ReflectionException
+     * Tests the holiday defined in this test before establishment.
+     * @throws \ReflectionException
      */
-    public function testStAndrewDayBefore2012()
+    public function testHolidayBeforeEstablishment()
     {
         $this->assertNotHoliday(
             self::REGION,
@@ -65,7 +63,7 @@ class StAndrewDayTest extends RomaniaBaseTestCase implements YasumiTestCaseInter
 
     /**
      * Tests the translated name of the holiday defined in this test.
-     * @throws ReflectionException
+     * @throws \ReflectionException
      */
     public function testTranslation(): void
     {
@@ -73,13 +71,13 @@ class StAndrewDayTest extends RomaniaBaseTestCase implements YasumiTestCaseInter
             self::REGION,
             self::HOLIDAY,
             $this->generateRandomYear(self::ESTABLISHMENT_YEAR),
-            [self::LOCALE => 'Sfântul Andrei']
+            [self::LOCALE => 'Spring Bank Holiday']
         );
     }
 
     /**
      * Tests type of the holiday defined in this test.
-     * @throws ReflectionException
+     * @throws \ReflectionException
      */
     public function testHolidayType(): void
     {
@@ -87,7 +85,7 @@ class StAndrewDayTest extends RomaniaBaseTestCase implements YasumiTestCaseInter
             self::REGION,
             self::HOLIDAY,
             $this->generateRandomYear(self::ESTABLISHMENT_YEAR),
-            Holiday::TYPE_OFFICIAL
+            Holiday::TYPE_BANK
         );
     }
 }